Fix typos

+++ Client/game_sa/CStreamingSA.cpp
@@ -136,7 +136,7 @@ void CStreamingSA::RequestSpecialModel(DWORD model, const char* szTexture, DWORD
     }
 }
 
-CStreamingInfo* CStreamingSA::GetStreamingInfoFromModelId(uint id)
+CStreamingInfo* CStreamingSA::GetStreamingInfoFromModelId(ushort id)
 {
     CStreamingInfo* pItemInfo = (CStreamingInfo*)(ARRAY_StreamModelInfo);
     return pItemInfo + id;
